:recycle: [Refactor]: 카테고리 색상 \& 알림 텍스트 찾는 함수 -\> 상수로 변경
