+++ src/entities/schedule/consts/index.ts
@@ -1,0 +1,2 @@
+export { CATEGORY_COLOR_MAP, getCategoryColorClass } from "./categoryColors";
+export { getNotificationText, NOTIFICATION_TEXT_MAP } from "./notificationText";
