--- src/entities/schedule/lib/index.ts
@@ -1,2 +1,0 @@
-export * from "./getCategoryColorClass";
-export * from "./getNotificationText";
